Add log rotation scripts
Logs are written to 'vault/logs/audit.log' as part of the Vault
auditing, when clients query the REST endpoints. The endpoint
'/get_status' in particular is exceptionally verbose, with a JSON field
containing a map of cumulative epochs to available key bits.

In a long running system, the total number of epochs accumulate,
resulting in a long log line per '/get_status' query. In a particular
production environment, a logging rate of 43 kBps was observed, with 580
stored epochs and opportunistic client queries to said endpoint every
second. This eventually resulted in the failure of the REST webservice
since the logs share the same partition as the Docker containers.

Log rotation can be performed to mitigate this, with this commit
supplying a default configuration for quick cron activation.
User-intervention is needed, since the default configuration will (1)
drop some auditing logs due to 'copytruncate', and (2) drop logs after
exceeding a size limit (currently set at 5x100MB) if no log backup is
set up. An alternative is to symlink the log file/directory to another
partition.